+++ api/src/main/java/org/allaymc/api/item/interfaces/ItemCrossbowStack.java
@@ -1,6 +1,7 @@
 package org.allaymc.api.item.interfaces;
 
 import org.allaymc.api.item.ItemStack;
+import org.allaymc.api.item.component.ItemCrossbowBaseComponent;
 
-public interface ItemCrossbowStack extends ItemStack {
+public interface ItemCrossbowStack extends ItemStack, ItemCrossbowBaseComponent {
 }